backend/agentcore/requirements.txt
bedrock-agentcore
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
requests>=2.31.0
